requirements.txt
httpx==0.28.1
python-dotenv==1.0.1
python-telegram-bot==21.10

